Added parameter to define prediciton extension and added pagexml input text equiv index as train parameter
