+++ calamari_ocr/ocr/datasets/dataset_factory.py
@@ -7,52 +7,8 @@ from .abbyy_dataset import AbbyyDataSet
 from .pagexml_dataset import PageXMLDataset
 from .hdf5_dataset import Hdf5DataSet
 from calamari_ocr.utils import keep_files_with_same_file_name
+from .datasetype import DataSetType
 
-
-class DataSetType(Enum):
-    RAW = 0
-    FILE = 1
-    ABBYY = 2
-    PAGEXML = 3
-    HDF5 = 4
-    EXTENDED_PREDICTION = 5
-    GENERATED_LINE = 6
-
-    def __str__(self):
-        return self.name
-
-    @staticmethod
-    def from_string(s):
-        try:
-            return DataSetType[s]
-        except KeyError:
-            raise ValueError()
-
-    @staticmethod
-    def files(type):
-        files_meta = {
-            DataSetType.RAW: False,
-            DataSetType.FILE: True,
-            DataSetType.ABBYY: True,
-            DataSetType.PAGEXML: True,
-            DataSetType.EXTENDED_PREDICTION: True,
-            DataSetType.HDF5: False,
-            DataSetType.GENERATED_LINE: False,
-        }
-
-        return files_meta[type]
-
-    @staticmethod
-    def gt_extension(type):
-        return {
-            DataSetType.RAW: None,
-            DataSetType.FILE: ".gt.txt",
-            DataSetType.ABBYY: ".abbyy.xml",
-            DataSetType.PAGEXML: ".xml",
-            DataSetType.EXTENDED_PREDICTION: ".json",
-            DataSetType.HDF5: ".h5",
-            DataSetType.GENERATED_LINE: None,
-        }[type]
 
 
 def create_dataset(type: DataSetType,
